[refactor]Reconstructing explicit memory
