+++ api/app/schemas/memory_episodic_schema.py
@@ -1,5 +1,5 @@
 """
-用户记忆相关的请求和响应模型
+情景记忆的请求和响应模型
 """
 from pydantic import BaseModel, Field
 from typing import Optional
@@ -28,16 +28,3 @@ class EpisodicMemoryDetailsRequest(BaseModel):
     
     end_user_id: str = Field(..., description="终端用户ID")
     summary_id: str = Field(..., description="情景记忆摘要ID")
-
-
-class ExplicitMemoryOverviewRequest(BaseModel):
-    """显性记忆总览查询请求"""
-    
-    end_user_id: str = Field(..., description="终端用户ID")
-
-
-class ExplicitMemoryDetailsRequest(BaseModel):
-    """显性记忆详情查询请求"""
-    
-    end_user_id: str = Field(..., description="终端用户ID")
-    memory_id: str = Field(..., description="记忆ID（情景记忆或语义记忆的ID）")
